major reworking for modularity and cleaning
