+++ pointillizer.com/myproject/myproject/myapp/templates/info.html
@@ -41,7 +41,7 @@
         <p> Some other <a href="https://photos.app.goo.gl/Dv6IObEJnsxKI3bn1">examples</a>, including gifs</p> 
         <p> Check out an experimental feature for gifs <a href={% url 'gif' guid_id=guid_id %}>here</a>, a little slow currently.
          <!-- Contacts and links  -->
-            <p><a href="https://github.com/tsennott/etc_projects/tree/master/Pointilism">github</a><br>
+            <p><a href="https://github.com/tsennott/pointillism">github</a><br>
             <script type="text/javascript">
             var first = "timothy.sennott";
             var last = "gmail.com";
